% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_operations.R
\name{athena_get_work_group}
\alias{athena_get_work_group}
\title{Returns information about the workgroup with the specified name}
\usage{
athena_get_work_group(WorkGroup)
}
\arguments{
\item{WorkGroup}{[required] The name of the workgroup.}
}
\description{
Returns information about the workgroup with the specified name.
}
\section{Request syntax}{
\preformatted{athena$get_work_group(
  WorkGroup = "string"
)
}
}

\keyword{internal}
